Remove stray funcs for parsers (#47) ~ Remove stray exported function for Roblox.Web.Parsers and move to WebParsers.ts -- https://git.mfdlabs.local/MFDLABS/RobloxLabsTypescript/issues/47#issuecomment-864618262
~ Remove all refs to SanitizeData --
https://git.mfdlabs.local/MFDLABS/RobloxLabsTypescript/issues/47#issuecomment-864618263
