+++ Source/ApiSites/Roblox.ClientSettingsCDN.Api/Controllers/v1/settings/application.ts
@@ -27,14 +27,14 @@
 
 import { Request, Response } from 'express';
 import { ClientSettings } from '../../../../../Assemblies/Platform/ClientSettings/Roblox.Platform.ClientSettings/Implementation/ClientSettingsUtil';
-import { SanitizeData } from '../../../../../Assemblies/Web/Parsers/Roblox.Web.Parsers/SanitizeData';
+import { WebParsers } from '../../../../../Assemblies/Web/Parsers/Roblox.Web.Parsers/WebParsers';
 import { DFLog, FASTLOGS } from '../../../../../Assemblies/Web/Util/Roblox.Web.Util/Logging/FastLog';
 import { InputValidator } from '../../../../../Assemblies/Web/Util/Roblox.Web.Util/Validators/InputValidator';
 
 export default {
 	method: 'all',
 	func: (request: Request, response: Response) => {
-		const group = SanitizeData(request.query.applicationName);
+		const group = WebParsers.SanitizeData(request.query.applicationName);
 		const FSettings = ClientSettings.GetFSettings();
 		FASTLOGS(DFLog('Tasks'), '[DFLog::Tasks] ClientSettings QuietGet Settings got group %s.', group);
