--- Source/Assemblies/Web/Parsers/Roblox.Web.Parsers/SanitizeData.ts
@@ -1,4 +1,0 @@
-export function SanitizeData(data: any) {
-	if (typeof data !== 'string') return null;
-	return (data || '').replace(/[^a-z0-9+]+/gi, '');
-}
