--- Source/Assemblies/Web/Parsers/Roblox.Web.Parsers/StripTheTrailingSlash.ts
@@ -1,3 +1,0 @@
-export function StripTheTrailingSlash(str: string) {
-	return str.replace(/\/$/, '');
-}
